Create class to enforce semantic versioning
Allow bypass of version bump in dev branch since it does not make sense
for the pipeline to decide arbitrarily on the next release version numer

BREAKING CHANGE: Build numbers appended to the version is considered
build meta-data so the seperator symbol is changing to the correct one
for this use. Published artifacts using '-' will not be recognised for
deploy-only jobs.